chore: hoist context to root
