+++ components/custom/survey/graphs/AffordFairhold.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import SurveyGraphCard from "@/components/custom/survey/SurveyGraphCard";
 import { PieChart, Pie, Legend, ResponsiveContainer, Cell } from "recharts";
-import { useSurveyContext } from "../../context";
-import { AFFORD_FAIRHOLD } from "../../constants";
+import { useSurveyContext } from "@context/surveyContext";
+import { AFFORD_FAIRHOLD } from "@lib/survey/constants";
 
 export const AffordFairhold = () => {
     let { affordFairhold } = useSurveyContext().barOrPie;
