+++ tsconfig.json
@@ -28,6 +28,7 @@
         "./*"
       ],
       "@components/*": ["components/*"],
+      "@context/*": ["context/*"],
       "@data/*": ["data/*"],
       "@hooks/*": ["hooks/*"],
       "@lib/*": ["lib/*"],
